--- test/A.py
@@ -1,4 +1,0 @@
-__author__ = 'Ben Christenson'
-__date__ = "1/5/18"
-
-import B
