--- test/C.py
@@ -1,5 +1,0 @@
-__author__ = 'Ben Christenson'
-__date__ = "1/5/18"
-
-print("hello, world!")
-raise BaseException('?')
